// src/types/enums.ts
export enum FaceEnum {
  Upper = 'UPPER',
  Left = 'LEFT',
  Front = 'FRONT',
  Right = 'RIGHT',
  Bottom = 'BOTTOM',
  Back = 'BACK',
}

export enum ColorsEnum {
  Yellow = 'rgb(250 204 21)',
  Blue = 'rgb(29 78 216)',
  Red = 'rgb(185 28 28)',
  Green = 'rgb(21 128 61)',
  White = 'rgb(255 255 255)',
  Orange = 'rgb(249 115 22)',
}